+++ api/src/main/java/org/allaymc/api/world/gamerule/GameRule.java
@@ -46,7 +46,8 @@ public enum GameRule {
     SHOW_DAYS_PLAYED("showDaysPlayed", false, Type.BOOLEAN),
     FUNCTION_COMMAND_LIMIT("functionCommandLimit", 10000, Type.INT),
     DO_LIMITED_CRAFTING("dolimitedcrafting", false, Type.BOOLEAN),
-    LOCATOR_BAR("locatorBar", true, Type.BOOLEAN);
+    LOCATOR_BAR("locatorBar", true, Type.BOOLEAN),
+    PLAYERS_SLEEPING_PERCENTAGE("playersSleepingPercentage", 100, Type.INT);
 
     private final String name;
     private final Object defaultValue;